+++ Example/QRCodeReader.swift/ViewController.swift
@@ -24,59 +24,92 @@
  *
  */
 
-import UIKit
 import AVFoundation
+import UIKit
 
 class ViewController: UIViewController, QRCodeReaderViewControllerDelegate {
-  lazy var reader: QRCodeReaderViewController = {
+  @IBOutlet weak var previewView: QRCodeReaderView! {
+    didSet {
+      previewView.setupComponents(with: QRCodeReaderViewControllerBuilder {
+        $0.reader                 = reader
+        $0.showTorchButton        = false
+        $0.showSwitchCameraButton = false
+        $0.showCancelButton       = false
+        $0.showOverlayView        = true
+        $0.rectOfInterest         = CGRect(x: 0.2, y: 0.2, width: 0.6, height: 0.6)
+      })
+    }
+  }
+  lazy var reader: QRCodeReader = QRCodeReader()
+  lazy var readerVC: QRCodeReaderViewController = {
     let builder = QRCodeReaderViewControllerBuilder {
-      $0.reader = QRCodeReader(metadataObjectTypes: [AVMetadataObjectTypeQRCode], captureDevicePosition: .back)
-      $0.showTorchButton = true
+      $0.reader                  = QRCodeReader(metadataObjectTypes: [.qr], captureDevicePosition: .back)
+      $0.showTorchButton         = true
+      $0.preferredStatusBarStyle = .lightContent
+      $0.showOverlayView         = true
+      $0.rectOfInterest          = CGRect(x: 0.2, y: 0.2, width: 0.6, height: 0.6)
+      
+      $0.reader.stopScanningWhenCodeIsFound = false
     }
     
     return QRCodeReaderViewController(builder: builder)
   }()
 
-  @IBAction func scanAction(_ sender: AnyObject) {
+  // MARK: - Actions
+
+  private func checkScanPermissions() -> Bool {
     do {
-      if try QRCodeReader.supportsMetadataObjectTypes() {
-        reader.modalPresentationStyle = .formSheet
-        reader.delegate               = self
-
-        reader.completionBlock = { (result: QRCodeReaderResult?) in
-          if let result = result {
-            print("Completion with result: \(result.value) of type \(result.metadataType)")
-          }
-        }
-
-        present(reader, animated: true, completion: nil)
-      }
+      return try QRCodeReader.supportsMetadataObjectTypes()
     } catch let error as NSError {
+      let alert: UIAlertController
+
       switch error.code {
       case -11852:
-        let alert = UIAlertController(title: "Error", message: "This app is not authorized to use Back Camera.", preferredStyle: .alert)
+        alert = UIAlertController(title: "Error", message: "This app is not authorized to use Back Camera.", preferredStyle: .alert)
 
         alert.addAction(UIAlertAction(title: "Setting", style: .default, handler: { (_) in
           DispatchQueue.main.async {
-            if let settingsURL = URL(string: UIApplicationOpenSettingsURLString) {
+            if let settingsURL = URL(string: UIApplication.openSettingsURLString) {
               UIApplication.shared.openURL(settingsURL)
             }
           }
         }))
+
         alert.addAction(UIAlertAction(title: "Cancel", style: .cancel, handler: nil))
-        present(alert, animated: true, completion: nil)
-
-
-
-      case -11814:
-        let alert = UIAlertController(title: "Error", message: "Reader not supported by the current device", preferredStyle: .alert)
+      default:
+        alert = UIAlertController(title: "Error", message: "Reader not supported by the current device", preferredStyle: .alert)
         alert.addAction(UIAlertAction(title: "OK", style: .cancel, handler: nil))
+      }
 
-        present(alert, animated: true, completion: nil)
-      default:()
+      present(alert, animated: true, completion: nil)
+
+      return false
+    }
+  }
+
+  @IBAction func scanInModalAction(_ sender: AnyObject) {
+    guard checkScanPermissions() else { return }
+
+    readerVC.modalPresentationStyle = .formSheet
+    readerVC.delegate               = self
+
+    readerVC.completionBlock = { (result: QRCodeReaderResult?) in
+      if let result = result {
+        print("Completion with result: \(result.value) of type \(result.metadataType)")
       }
     }
 
+    present(readerVC, animated: true, completion: nil)
+  }
+
+  @IBAction func scanInPreviewAction(_ sender: Any) {
+    guard checkScanPermissions(), !reader.isRunning else { return }
+
+    reader.didFindCode = { result in
+      print("Completion with result: \(result.value) of type \(result.metadataType)")
+    }
+
+    reader.startScanning()
   }
 
   // MARK: - QRCodeReader Delegate Methods
@@ -97,9 +130,7 @@ class ViewController: UIViewController, QRCodeReaderViewControllerDelegate {
   }
 
   func reader(_ reader: QRCodeReaderViewController, didSwitchCamera newCaptureDevice: AVCaptureDeviceInput) {
-    if let cameraName = newCaptureDevice.device.localizedName {
-      print("Switching capturing to: \(cameraName)")
-    }
+    print("Switching capture to: \(newCaptureDevice.device.localizedName)")
   }
 
   func readerDidCancel(_ reader: QRCodeReaderViewController) {