+++ Sources/QRCodeReader.swift
@@ -27,17 +27,25 @@
 import UIKit
 import AVFoundation
 
+protocol QRCodeReaderLifeCycleDelegate: class {
+  func readerDidStartScanning()
+  func readerDidStopScanning()
+}
+
 /// Reader object base on the `AVCaptureDevice` to read / scan 1D and 2D codes.
 public final class QRCodeReader: NSObject, AVCaptureMetadataOutputObjectsDelegate {
-  var defaultDevice: AVCaptureDevice = .defaultDevice(withMediaType: AVMediaTypeVideo)
-  var frontDevice: AVCaptureDevice?  = {
+  private let sessionQueue         = DispatchQueue(label: "session queue")
+  private let metadataObjectsQueue = DispatchQueue(label: "com.yannickloriot.qr", attributes: [], target: nil)
+  
+  let defaultDevice: AVCaptureDevice? = AVCaptureDevice.default(for: .video)
+  let frontDevice: AVCaptureDevice?   = {
     if #available(iOS 10, *) {
-      return AVCaptureDevice.defaultDevice(withDeviceType: .builtInWideAngleCamera, mediaType: AVMediaTypeVideo, position: .front)
+      return AVCaptureDevice.default(.builtInWideAngleCamera, for: AVMediaType.video, position: .front)
     }
     else {
-      for device in AVCaptureDevice.devices(withMediaType: AVMediaTypeVideo) {
-        if let _device = device as? AVCaptureDevice , _device.position == AVCaptureDevicePosition.front {
-          return _device
+      for device in AVCaptureDevice.devices(for: AVMediaType.video) {
+        if device.position == .front {
+          return device
         }
       }
     }
@@ -46,7 +54,9 @@ public final class QRCodeReader: NSObject, AVCaptureMetadataOutputObjectsDelegat
   }()
 
   lazy var defaultDeviceInput: AVCaptureDeviceInput? = {
-    return try? AVCaptureDeviceInput(device: self.defaultDevice)
+    guard let defaultDevice = defaultDevice else { return nil }
+
+    return try? AVCaptureDeviceInput(device: defaultDevice)
   }()
 
   lazy var frontDeviceInput: AVCaptureDeviceInput? = {
@@ -57,18 +67,18 @@ public final class QRCodeReader: NSObject, AVCaptureMetadataOutputObjectsDelegat
     return nil
   }()
 
-  var metadataOutput = AVCaptureMetadataOutput()
-  var session        = AVCaptureSession()
+  let session        = AVCaptureSession()
+  let metadataOutput = AVCaptureMetadataOutput()
+
+  weak var lifeCycleDelegate: QRCodeReaderLifeCycleDelegate?
 
   // MARK: - Managing the Properties
 
   /// CALayer that you use to display video as it is being captured by an input device.
-  public lazy var previewLayer: AVCaptureVideoPreviewLayer = {
-    return AVCaptureVideoPreviewLayer(session: self.session)
-  }()
+  public let previewLayer: AVCaptureVideoPreviewLayer
 
-  /// An array of strings identifying the types of metadata objects to process.
-  public let metadataObjectTypes: [String]
+  /// An array of object identifying the types of metadata objects to process.
+  public let metadataObjectTypes: [AVMetadataObject.ObjectType]
 
   // MARK: - Managing the Code Discovery
 
@@ -79,7 +89,7 @@ public final class QRCodeReader: NSObject, AVCaptureMetadataOutputObjectsDelegat
   public var didFindCode: ((QRCodeReaderResult) -> Void)?
 
   /// Block is executed when a found metadata object string could not be decoded.
-  public var didFailDecoding: ((Void) -> Void)?
+  public var didFailDecoding: (() -> Void)?
 
   // MARK: - Creating the Code Reade
 
@@ -87,15 +97,15 @@ public final class QRCodeReader: NSObject, AVCaptureMetadataOutputObjectsDelegat
    Initializes the code reader with the QRCode metadata type object.
    */
   public convenience override init() {
-    self.init(metadataObjectTypes: [AVMetadataObjectTypeQRCode], captureDevicePosition: .back)
+    self.init(metadataObjectTypes: [AVMetadataObject.ObjectType.qr], captureDevicePosition: .back)
   }
 
   /**
    Initializes the code reader with an array of metadata object types, and the default initial capture position
 
-   - parameter metadataObjectTypes: An array of strings identifying the types of metadata objects to process.
+   - parameter metadataObjectTypes: An array of objects identifying the types of metadata objects to process.
    */
-  public convenience init(metadataObjectTypes types: [String]) {
+  public convenience init(metadataObjectTypes types: [AVMetadataObject.ObjectType]) {
     self.init(metadataObjectTypes: types, captureDevicePosition: .back)
   }
 
@@ -104,44 +114,62 @@ public final class QRCodeReader: NSObject, AVCaptureMetadataOutputObjectsDelegat
 
    - parameter captureDevicePosition: The capture position to use on start of scanning
    */
-  public convenience init(captureDevicePosition position: AVCaptureDevicePosition) {
-    self.init(metadataObjectTypes: [AVMetadataObjectTypeQRCode], captureDevicePosition: position)
+  public convenience init(captureDevicePosition position: AVCaptureDevice.Position) {
+    self.init(metadataObjectTypes: [AVMetadataObject.ObjectType.qr], captureDevicePosition: position)
   }
 
   /**
    Initializes the code reader with an array of metadata object types.
 
-   - parameter metadataObjectTypes: An array of strings identifying the types of metadata objects to process.
+   - parameter metadataObjectTypes: An array of objects identifying the types of metadata objects to process.
    - parameter captureDevicePosition: The Camera to use on start of scanning.
    */
-  public init(metadataObjectTypes types: [String], captureDevicePosition: AVCaptureDevicePosition) {
+  public init(metadataObjectTypes types: [AVMetadataObject.ObjectType], captureDevicePosition: AVCaptureDevice.Position) {
     metadataObjectTypes = types
+    previewLayer        = AVCaptureVideoPreviewLayer(session: session)
 
     super.init()
 
-    configureDefaultComponents(withCaptureDevicePosition: captureDevicePosition)
+    sessionQueue.async {
+      self.configureDefaultComponents(withCaptureDevicePosition: captureDevicePosition)
+    }
   }
 
   // MARK: - Initializing the AV Components
 
-  private func configureDefaultComponents(withCaptureDevicePosition: AVCaptureDevicePosition) {
-    session.addOutput(metadataOutput)
+  private func configureDefaultComponents(withCaptureDevicePosition: AVCaptureDevice.Position) {
+    for output in session.outputs {
+      session.removeOutput(output)
+    }
+    for input in session.inputs {
+      session.removeInput(input)
+    }
 
+    // Add video input
     switch withCaptureDevicePosition {
     case .front:
       if let _frontDeviceInput = frontDeviceInput {
         session.addInput(_frontDeviceInput)
       }
-    case .back, .unspecified:
+    default:
       if let _defaultDeviceInput = defaultDeviceInput {
         session.addInput(_defaultDeviceInput)
       }
     }
 
+    // Add metadata output
+    session.addOutput(metadataOutput)
+    metadataOutput.setMetadataObjectsDelegate(self, queue: metadataObjectsQueue)
 
-    metadataOutput.setMetadataObjectsDelegate(self, queue: DispatchQueue.main)
-    metadataOutput.metadataObjectTypes = metadataObjectTypes
-    previewLayer.videoGravity          = AVLayerVideoGravityResizeAspectFill
+    let allTypes = Set(metadataOutput.availableMetadataObjectTypes)
+    let filtered = metadataObjectTypes.filter { (mediaType) -> Bool in
+      allTypes.contains(mediaType)
+    }
+
+    metadataOutput.metadataObjectTypes = filtered
+    previewLayer.videoGravity          = .resizeAspectFill
+
+    session.commitConfiguration()
   }
 
   /// Switch between the back and the front camera.
@@ -154,11 +182,12 @@ public final class QRCodeReader: NSObject, AVCaptureMetadataOutputObjectsDelegat
         session.removeInput(_currentInput)
 
         let newDeviceInput = (_currentInput.device.position == .front) ? defaultDeviceInput : _frontDeviceInput
-        session.addInput(newDeviceInput)
+        session.addInput(newDeviceInput!)
       }
 
       session.commitConfiguration()
     }
+
     return session.inputs.first as? AVCaptureDeviceInput
   }
 
@@ -170,15 +199,27 @@ public final class QRCodeReader: NSObject, AVCaptureMetadataOutputObjectsDelegat
    *Notes: if `stopScanningWhenCodeIsFound` is sets to true (default behaviour), each time the scanner found a code it calls the `stopScanning` method.*
    */
   public func startScanning() {
-    if !session.isRunning {
-      session.startRunning()
+    sessionQueue.async {
+      guard !self.session.isRunning else { return }
+
+      self.session.startRunning()
+
+      DispatchQueue.main.async {
+        self.lifeCycleDelegate?.readerDidStartScanning()
+      }
     }
   }
 
   /// Stops scanning the codes.
   public func stopScanning() {
-    if session.isRunning {
-      session.stopRunning()
+    sessionQueue.async {
+      guard self.session.isRunning else { return }
+
+      self.session.stopRunning()
+
+      DispatchQueue.main.async {
+        self.lifeCycleDelegate?.readerDidStopScanning()
+      }
     }
   }
 
@@ -208,7 +249,7 @@ public final class QRCodeReader: NSObject, AVCaptureMetadataOutputObjectsDelegat
    - returns: true if a torch is available.
    */
   public var isTorchAvailable: Bool {
-    return defaultDevice.isTorchAvailable
+    return defaultDevice?.isTorchAvailable ?? false
   }
 
   /**
@@ -216,12 +257,11 @@ public final class QRCodeReader: NSObject, AVCaptureMetadataOutputObjectsDelegat
    */
   public func toggleTorch() {
     do {
-      try defaultDevice.lockForConfiguration()
+      try defaultDevice?.lockForConfiguration()
 
-      let current             = defaultDevice.torchMode
-      defaultDevice.torchMode = AVCaptureTorchMode.on == current ? .off : .on
+      defaultDevice?.torchMode = defaultDevice?.torchMode == .on ? .off : .on
 
-      defaultDevice.unlockForConfiguration()
+      defaultDevice?.unlockForConfiguration()
     }
     catch _ { }
   }
@@ -279,10 +319,10 @@ public final class QRCodeReader: NSObject, AVCaptureMetadataOutputObjectsDelegat
       return .landscapeLeft
     case .landscapeRight:
       return .landscapeRight
-    case .portrait:
-      return .portrait
     case .portraitUpsideDown:
       return .portraitUpsideDown
+    default:
+      return .portrait
     }
   }
 
@@ -294,7 +334,7 @@ public final class QRCodeReader: NSObject, AVCaptureMetadataOutputObjectsDelegat
    - returns: A boolean value that indicates whether the reader is available.
    */
   public class func isAvailable() -> Bool {
-    let captureDevice = AVCaptureDevice.defaultDevice(withMediaType: AVMediaTypeVideo)
+    guard let captureDevice = AVCaptureDevice.default(for: .video) else { return false }
 
     return (try? AVCaptureDeviceInput(device: captureDevice)) != nil
   }
@@ -302,17 +342,18 @@ public final class QRCodeReader: NSObject, AVCaptureMetadataOutputObjectsDelegat
   /**
    Checks and return whether the given metadata object types are supported by the current device.
 
-   - parameter metadataTypes: An array of strings identifying the types of metadata objects to check.
+   - parameter metadataTypes: An array of objects identifying the types of metadata objects to check.
 
    - returns: A boolean value that indicates whether the device supports the given metadata object types.
    */
-  public class func supportsMetadataObjectTypes(_ metadataTypes: [String]? = nil) throws -> Bool {
-    let captureDevice = AVCaptureDevice.defaultDevice(withMediaType: AVMediaTypeVideo)
+  public class func supportsMetadataObjectTypes(_ metadataTypes: [AVMetadataObject.ObjectType]? = nil) throws -> Bool {
+    guard let captureDevice = AVCaptureDevice.default(for: .video) else {
+      throw NSError(domain: "com.yannickloriot.error", code: -1001, userInfo: nil)
+    }
 
     let deviceInput = try AVCaptureDeviceInput(device: captureDevice)
-
-    let output  = AVCaptureMetadataOutput()
-    let session = AVCaptureSession()
+    let output      = AVCaptureMetadataOutput()
+    let session     = AVCaptureSession()
 
     session.addInput(deviceInput)
     session.addOutput(output)
@@ -321,11 +362,11 @@ public final class QRCodeReader: NSObject, AVCaptureMetadataOutputObjectsDelegat
 
     if metadataObjectTypes == nil || metadataObjectTypes?.count == 0 {
       // Check the QRCode metadata object type by default
-      metadataObjectTypes = [AVMetadataObjectTypeQRCode]
+      metadataObjectTypes = [.qr]
     }
 
     for metadataObjectType in metadataObjectTypes! {
-      if !output.availableMetadataObjectTypes.contains(where: { $0 as! String == metadataObjectType }) {
+      if !output.availableMetadataObjectTypes.contains { $0 == metadataObjectType } {
         return false
       }
     }
@@ -335,26 +376,34 @@ public final class QRCodeReader: NSObject, AVCaptureMetadataOutputObjectsDelegat
 
   // MARK: - AVCaptureMetadataOutputObjects Delegate Methods
 
-  public func captureOutput(_ captureOutput: AVCaptureOutput!, didOutputMetadataObjects metadataObjects: [Any]!, from connection: AVCaptureConnection!) {
-    for current in metadataObjects {
-      if let _readableCodeObject = current as? AVMetadataMachineReadableCodeObject {
-        if _readableCodeObject.stringValue != nil {
-          if metadataObjectTypes.contains(_readableCodeObject.type) {
-            if let sVal = _readableCodeObject.stringValue {
-              if stopScanningWhenCodeIsFound {
-                stopScanning()
+  public func metadataOutput(_ output: AVCaptureMetadataOutput, didOutput metadataObjects: [AVMetadataObject], from connection: AVCaptureConnection) {
+    sessionQueue.async { [weak self] in
+      guard let weakSelf = self else { return }
+
+      for current in metadataObjects {
+        if let _readableCodeObject = current as? AVMetadataMachineReadableCodeObject {
+          if _readableCodeObject.stringValue != nil {
+            if weakSelf.metadataObjectTypes.contains(_readableCodeObject.type) {
+              guard weakSelf.session.isRunning, let sVal = _readableCodeObject.stringValue else { return }
+
+              if weakSelf.stopScanningWhenCodeIsFound {
+                weakSelf.session.stopRunning()
+
+                DispatchQueue.main.async {
+                  weakSelf.lifeCycleDelegate?.readerDidStopScanning()
+                }
               }
 
-              let scannedResult = QRCodeReaderResult(value: sVal, metadataType:_readableCodeObject.type)
+              let scannedResult = QRCodeReaderResult(value: sVal, metadataType:_readableCodeObject.type.rawValue)
 
-              DispatchQueue.main.async(execute: { [weak self] in
-                self?.didFindCode?(scannedResult)
-              })
+              DispatchQueue.main.async {
+                weakSelf.didFindCode?(scannedResult)
+              }
             }
           }
         }
         else {
-          didFailDecoding?()
+          weakSelf.didFailDecoding?()
         }
       }
     }