+++ Sources/QRCodeReaderViewController.swift
@@ -29,15 +29,12 @@ import AVFoundation
 
 /// Convenient controller to display a view to scan/read 1D or 2D bar codes like the QRCodes. It is based on the `AVFoundation` framework from Apple. It aims to replace ZXing or ZBar for iOS 7 and over.
 public class QRCodeReaderViewController: UIViewController {
-  /// The code reader object used to scan the bar code.
-  public let codeReader: QRCodeReader
+  private let builder: QRCodeReaderViewControllerBuilder
 
-  let readerView: QRCodeReaderContainer
-  let startScanningAtLoad: Bool
-  let showCancelButton: Bool
-  let showSwitchCameraButton: Bool
-  let showTorchButton: Bool
-  let showOverlayView: Bool
+  /// The code reader object used to scan the bar code.
+  public var codeReader: QRCodeReader {
+    return builder.reader
+  }
 
   // MARK: - Managing the Callback Responders
 
@@ -61,13 +58,7 @@ public class QRCodeReaderViewController: UIViewController {
    - parameter builder: A QRCodeViewController builder object.
    */
   required public init(builder: QRCodeReaderViewControllerBuilder) {
-    readerView             = builder.readerView
-    startScanningAtLoad    = builder.startScanningAtLoad
-    codeReader             = builder.reader
-    showCancelButton       = builder.showCancelButton
-    showSwitchCameraButton = builder.showSwitchCameraButton
-    showTorchButton        = builder.showTorchButton
-    showOverlayView        = builder.showOverlayView
+    self.builder = builder
 
     super.init(nibName: nil, bundle: nil)
 
@@ -75,7 +66,7 @@ public class QRCodeReaderViewController: UIViewController {
 
     codeReader.didFindCode = { [weak self] resultAsObject in
       if let weakSelf = self {
-        if let qrv = weakSelf.readerView.displayable as? QRCodeReaderView {
+        if let qrv = builder.readerView.displayable as? QRCodeReaderView {
           qrv.addGreenBorder()
         }
         weakSelf.completionBlock?(resultAsObject)
@@ -83,40 +74,29 @@ public class QRCodeReaderViewController: UIViewController {
       }
     }
 
-    codeReader.didFailDecoding = { [weak self] in
-      if let weakSelf = self {
-        if let qrv = weakSelf.readerView.displayable as? QRCodeReaderView {
-          qrv.addRedBorder()
-        }
+    codeReader.didFailDecoding = {
+      if let qrv = builder.readerView.displayable as? QRCodeReaderView {
+        qrv.addRedBorder()
       }
     }
 
     setupUIComponentsWithCancelButtonTitle(builder.cancelButtonTitle)
-
-    NotificationCenter.default.addObserver(self, selector: #selector(orientationDidChange), name: NSNotification.Name.UIDeviceOrientationDidChange, object: nil)
   }
 
   required public init?(coder aDecoder: NSCoder) {
-    codeReader             = QRCodeReader()
-    readerView             = QRCodeReaderContainer(displayable: QRCodeReaderView())
-    startScanningAtLoad    = false
-    showCancelButton       = false
-    showTorchButton        = false
-    showSwitchCameraButton = false
-    showOverlayView        = false
+    self.builder = QRCodeReaderViewControllerBuilder()
 
     super.init(coder: aDecoder)
   }
 
   // MARK: - Responding to View Events
-  override public var supportedInterfaceOrientations: UIInterfaceOrientationMask {
-    return parent?.supportedInterfaceOrientations ?? .all
-  }
 
   override public func viewWillAppear(_ animated: Bool) {
     super.viewWillAppear(animated)
 
-    if startScanningAtLoad {
+    if builder.startScanningAtLoad {
+      builder.readerView.displayable.setNeedsUpdateOrientation()
+
       startScanning()
     }
   }
@@ -133,48 +113,36 @@ public class QRCodeReaderViewController: UIViewController {
     codeReader.previewLayer.frame = view.bounds
   }
 
-  // MARK: - Managing the Orientation
-
-  func orientationDidChange(_ notification: Notification) {
-    readerView.view.setNeedsDisplay()
-
-    if let device = notification.object as? UIDevice , codeReader.previewLayer.connection.isVideoOrientationSupported {
-      codeReader.previewLayer.connection.videoOrientation = QRCodeReader.videoOrientation(deviceOrientation: device.orientation, withSupportedOrientations: supportedInterfaceOrientations, fallbackOrientation: codeReader.previewLayer.connection.videoOrientation)
-    }
+  public override var preferredStatusBarStyle: UIStatusBarStyle {
+    return builder.preferredStatusBarStyle ?? super.preferredStatusBarStyle
   }
 
   // MARK: - Initializing the AV Components
 
   private func setupUIComponentsWithCancelButtonTitle(_ cancelButtonTitle: String) {
-    view.addSubview(readerView.view)
+    view.addSubview(builder.readerView.view)
 
-    let sscb = showSwitchCameraButton && codeReader.hasFrontDevice
-    let stb  = showTorchButton && codeReader.isTorchAvailable
-
-    readerView.setupComponents(showCancelButton: showCancelButton, showSwitchCameraButton: sscb, showTorchButton: stb, showOverlayView: showOverlayView)
+    builder.readerView.view.translatesAutoresizingMaskIntoConstraints = false
+    builder.readerView.setupComponents(with: builder)
 
     // Setup action methods
 
-    readerView.displayable.switchCameraButton?.addTarget(self, action: #selector(switchCameraAction), for: .touchUpInside)
-    readerView.displayable.toggleTorchButton?.addTarget(self, action: #selector(toggleTorchAction), for: .touchUpInside)
-    readerView.displayable.cancelButton?.setTitle(cancelButtonTitle, for: .normal)
-    readerView.displayable.cancelButton?.addTarget(self, action: #selector(cancelAction), for: .touchUpInside)
-
-    // Setup camera preview layer
-    codeReader.previewLayer.frame = CGRect(x: 0, y: 0, width: view.frame.size.width, height: view.frame.size.height)
-
-    if codeReader.previewLayer.connection.isVideoOrientationSupported {
-      let orientation = UIDevice.current.orientation
-
-      codeReader.previewLayer.connection.videoOrientation = QRCodeReader.videoOrientation(deviceOrientation: orientation, withSupportedOrientations: supportedInterfaceOrientations)
-    }
-
-    readerView.displayable.cameraView.layer.insertSublayer(codeReader.previewLayer, at: 0)
+    builder.readerView.displayable.switchCameraButton?.addTarget(self, action: #selector(switchCameraAction), for: .touchUpInside)
+    builder.readerView.displayable.toggleTorchButton?.addTarget(self, action: #selector(toggleTorchAction), for: .touchUpInside)
+    builder.readerView.displayable.cancelButton?.setTitle(cancelButtonTitle, for: .normal)
+    builder.readerView.displayable.cancelButton?.addTarget(self, action: #selector(cancelAction), for: .touchUpInside)
 
     // Setup constraints
 
-    for attribute in [NSLayoutAttribute.left, NSLayoutAttribute.top, NSLayoutAttribute.right, NSLayoutAttribute.bottom] {
-      view.addConstraint(NSLayoutConstraint(item: readerView.view, attribute: attribute, relatedBy: .equal, toItem: view, attribute: attribute, multiplier: 1, constant: 0))
+    for attribute in [.left, .top, .right] as [NSLayoutConstraint.Attribute] {
+        NSLayoutConstraint(item: builder.readerView.view, attribute: attribute, relatedBy: .equal, toItem: view, attribute: attribute, multiplier: 1, constant: 0).isActive = true
+    }
+    
+    if #available(iOS 11.0, *) {
+        view.safeAreaLayoutGuide.bottomAnchor.constraint(equalTo: builder.readerView.view.bottomAnchor).isActive = true
+    }
+    else {
+        NSLayoutConstraint(item: builder.readerView.view, attribute: .bottom, relatedBy: .equal, toItem: view, attribute: .bottom, multiplier: 1, constant: 0).isActive = true
     }
   }
 
@@ -192,7 +160,7 @@ public class QRCodeReaderViewController: UIViewController {
 
   // MARK: - Catching Button Events
 
-  func cancelAction(_ button: UIButton) {
+  @objc func cancelAction(_ button: UIButton) {
     codeReader.stopScanning()
 
     if let _completionBlock = completionBlock {
@@ -202,13 +170,13 @@ public class QRCodeReaderViewController: UIViewController {
     delegate?.readerDidCancel(self)
   }
 
-  func switchCameraAction(_ button: SwitchCameraButton) {
+  @objc func switchCameraAction(_ button: SwitchCameraButton) {
     if let newDevice = codeReader.switchDeviceInput() {
       delegate?.reader(self, didSwitchCamera: newDevice)
     }
   }
   
-  func toggleTorchAction(_ button: ToggleTorchButton) {
+  @objc func toggleTorchAction(_ button: ToggleTorchButton) {
     codeReader.toggleTorch()
   }
 }