+++ Sources/QRCodeReaderViewControllerBuilder.swift
@@ -25,6 +25,7 @@
  */
 
 import Foundation
+import UIKit
 
 /**
  The QRCodeViewControllerBuilder aims to create a simple configuration object for
@@ -67,18 +68,60 @@ public final class QRCodeReaderViewControllerBuilder {
   /**
    Flag to display the switch camera button.
    */
-  public var showSwitchCameraButton = true
+  public var showSwitchCameraButton: Bool {
+    get {
+      return _showSwitchCameraButton && reader.hasFrontDevice
+    }
+    set {
+      _showSwitchCameraButton = newValue
+    }
+  }
+  private var _showSwitchCameraButton: Bool = true
 
   /**
    Flag to display the toggle torch button. If the value is true and there is no torch the button will not be displayed.
    */
-  public var showTorchButton = false
+  public var showTorchButton: Bool {
+    get {
+      return _showTorchButton && reader.isTorchAvailable
+    }
+    set {
+      _showTorchButton = newValue
+    }
+  }
+  private var _showTorchButton = true
 
   /**
    Flag to display the guide view.
    */
-  public var showOverlayView = true
+  public var showOverlayView = false
 
+  /**
+    Flag to display the guide view.
+  */
+  public var handleOrientationChange = true
+
+  /**
+   A UIStatusBarStyle key indicating your preferred status bar style for the view controller.
+   Nil by default. It means it'll use the current context status bar style.
+  */
+  public var preferredStatusBarStyle: UIStatusBarStyle? = nil
+
+  /**
+   Specifies a rectangle of interest for limiting the search area for visual metadata.
+
+   The value of this property is a CGRect that determines the receiver's rectangle of interest for each frame of video. The rectangle's origin is top left and is relative to the coordinate space of the device providing the metadata. Specifying a rectOfInterest may improve detection performance for certain types of metadata. The default value of this property is the value CGRectMake(0, 0, 1, 1). Metadata objects whose bounds do not intersect with the rectOfInterest will not be returned.
+  */
+  public var rectOfInterest: CGRect = CGRect(x: 0, y: 0, width: 1, height: 1) {
+    didSet {
+      reader.metadataOutput.rectOfInterest = CGRect(
+        x: min(max(rectOfInterest.origin.x, 0), 1),
+        y: min(max(rectOfInterest.origin.y, 0), 1),
+        width: min(max(rectOfInterest.width, 0), 1),
+        height: min(max(rectOfInterest.height, 0), 1)
+      )
+    }
+  }
 
   // MARK: - Initializing a Flap View