+++ Sources/ReaderOverlayView.swift
@@ -26,6 +26,22 @@
 
 import UIKit
 
+/// The overlay state
+public enum QRCodeReaderViewOverlayState {
+  /// The overlay is in normal state
+  case normal
+  /// The overlay is in valid state
+  case valid
+  /// The overlay is in wrong state
+  case wrong
+}
+
+/// The overlay protocol
+public protocol QRCodeReaderViewOverlay: UIView {
+  /// Set the state of the overlay
+  func setState(_ state: QRCodeReaderViewOverlayState)
+}
+
 /// Overlay over the camera view to display the area (a square) where to scan the code.
 public final class ReaderOverlayView: UIView {
   private var overlay: CAShapeLayer = {
@@ -39,6 +55,30 @@ public final class ReaderOverlayView: UIView {
 
     return overlay
   }()
+  
+  private var state: QRCodeReaderViewOverlayState = .normal {
+    didSet {
+      switch state {
+      case .normal:
+        overlay.strokeColor = defaultColor.cgColor
+      case .valid:
+        overlay.strokeColor = highlightValidColor.cgColor
+      case .wrong:
+        overlay.strokeColor = highlightWrongColor.cgColor
+      }
+      
+      setNeedsDisplay()
+    }
+  }
+  
+  /// The default overlay color
+  public var defaultColor: UIColor = .white
+  
+  /// The overlay color when a valid code has been scanned
+  public var highlightValidColor: UIColor = .green
+  
+  /// The overlay color when a wrong code has been scanned
+  public var highlightWrongColor: UIColor = .red
 
   override init(frame: CGRect) {
     super.init(frame: frame)
@@ -53,31 +93,31 @@ public final class ReaderOverlayView: UIView {
   }
 
   private func setupOverlay() {
+    state = .normal
+    
     layer.addSublayer(overlay)
   }
 
-  var overlayColor: UIColor = UIColor.white {
+  var rectOfInterest: CGRect = CGRect(x: 0, y: 0, width: 1, height: 1) {
     didSet {
-      self.overlay.strokeColor = overlayColor.cgColor
-      self.setNeedsDisplay()
+      setNeedsDisplay()
     }
   }
 
   public override func draw(_ rect: CGRect) {
-    var innerRect = rect.insetBy(dx: 50, dy: 50)
-    let minSize   = min(innerRect.width, innerRect.height)
+    let innerRect = CGRect(
+      x: rect.width * rectOfInterest.minX,
+      y: rect.height * rectOfInterest.minY,
+      width: rect.width * rectOfInterest.width,
+      height: rect.height * rectOfInterest.height
+    )
 
-    if innerRect.width != minSize {
-      innerRect.origin.x   += (innerRect.width - minSize) / 2
-      innerRect.size.width = minSize
-    }
-    else if innerRect.height != minSize {
-      innerRect.origin.y    += (innerRect.height - minSize) / 2
-      innerRect.size.height = minSize
-    }
+    overlay.path = UIBezierPath(roundedRect: innerRect, cornerRadius: 5).cgPath
+  }
+}
 
-    let offsetRect = innerRect.offsetBy(dx: 0, dy: 15)
-
-    overlay.path  = UIBezierPath(roundedRect: offsetRect, cornerRadius: 5).cgPath
+extension ReaderOverlayView: QRCodeReaderViewOverlay {
+  public func setState(_ state: QRCodeReaderViewOverlayState) {
+    self.state = state
   }
 }